To-dos can now be moved by drag and drop
Added a new field to the Todo entity which keeps track of it's current
position in the to-do list it belongs to. The Todo entity class now
implements the Comparable interface, so that a List containing to-dos
can be sorted easily. Because of the relation defined in ListWithTodos,
the order cannot be given in the database query and so the list must be
always sorted after fetching from the database.
Added a database migration reflecting this entity change.
The app module's gradle file dependencies got updated, and changes in
Workers are reflected in the worker classes.
The database migration test is currently broken, for reasons.
The migration should however work.